LayoutTests/fast/css/getComputedStyle/getComputedStyle-padding-margin-subpixel-length-expected.txt

Test calling getPropertyValue on computed margin styles.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getComputedStyle(testElement).marginTop is "1.5px"
PASS getComputedStyle(testElement).marginRight is "2.5px"
PASS getComputedStyle(testElement).marginBottom is "3.5px"
PASS getComputedStyle(testElement).marginLeft is "4.5px"
PASS getComputedStyle(testElement).margin is "1.5px 2.5px 3.5px 4.5px"
PASS getComputedStyle(testElement).paddingTop is "5.5px"
PASS getComputedStyle(testElement).paddingRight is "6.5px"
PASS getComputedStyle(testElement).paddingBottom is "7.5px"
PASS getComputedStyle(testElement).paddingLeft is "8.5px"
PASS getComputedStyle(testElement).padding is "5.5px 6.5px 7.5px 8.5px"
PASS successfullyParsed is true

TEST COMPLETE
